refactor: extract gateway status api surface
