+++ crates/daemon/src/gateway/control.rs
@@ -53,7 +53,10 @@ use super::pairing_runtime::{
 use super::read_models::{
     GatewayChannelInventoryReadModel, GatewayPairingSessionLeaseReadModel,
     GatewayRuntimeSnapshotReadModel,
-    build_gateway_operator_summary_from_registry_read_model,
+};
+use super::status_api::{
+    handle_gateway_channels, handle_gateway_operator_summary, handle_gateway_runtime_snapshot,
+    handle_gateway_status,
 };
 use super::state::{
     GatewayControlSurfaceBinding, gateway_control_token_path,
@@ -591,72 +594,6 @@ fn build_gateway_control_router(app_state: Arc<GatewayControlAppState>) -> Route
         .route("/v1/chat/completions", post(handle_chat_completions))
         .route("/health", get(handle_health))
         .with_state(app_state)
-}
-
-async fn handle_gateway_status(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    let status = match request.status() {
-        Ok(status) => status,
-        Err(response) => return response,
-    };
-    gateway_control_payload_response(&status, "gateway status payload")
-}
-
-async fn handle_gateway_channels(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    gateway_control_payload_response(
-        request.app_state().channel_inventory.as_ref(),
-        "gateway channels payload",
-    )
-}
-
-async fn handle_gateway_runtime_snapshot(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    gateway_control_payload_response(
-        request.app_state().runtime_snapshot.as_ref(),
-        "gateway runtime snapshot payload",
-    )
-}
-
-async fn handle_gateway_operator_summary(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    let status = match request.status() {
-        Ok(status) => status,
-        Err(response) => return response,
-    };
-    let pairing_registry = gateway_pairing_registry(request.app_state()).ok();
-    let summary = build_gateway_operator_summary_from_registry_read_model(
-        &status,
-        request.app_state().channel_inventory.as_ref(),
-        request.app_state().runtime_snapshot.as_ref(),
-        request.app_state().config_path.as_str(),
-        pairing_registry.as_ref(),
-    );
-    gateway_control_payload_response(&summary, "gateway operator summary payload")
 }
 
 async fn handle_gateway_stop(
